long dlightbits

+++ gl_light.c
@@ -256,7 +256,7 @@ rewritten to use LordHavoc's lighting speedup
 recursive goes through the nodes marking the surfaces near the dynamic light as lit
 =============
 */
-void R_MarkLights (dlight_t *light, int bit, mnode_t *node)
+void R_MarkLights (dlight_t *light, unsigned long long bit, mnode_t *node)
 {
 	mplane_t	*plane;
 	msurface_t	*surf;
@@ -357,7 +357,7 @@ void R_PushDlights (void)
 		if (l->die < cl.time || !l->radius)
 			continue;
 		
-		R_MarkLights (l, 1<<i, cl.worldmodel->nodes);
+		R_MarkLights (l, 1ULL<<i, cl.worldmodel->nodes);
 	}
 }
